OpenEMS Edge Simulator (and more...) (#1087)
This "Simulator App" for OpenEMS Edge provides a _simulated real-time_ environment. The main purpose is to be able to run a long period of input data on the OpenEMS stack and execute control algorithms just like they were executed in real-time, but with included 'fast-forward'.

For the moment it is possible e.g. to simulate around 3600 cycles (= about one hour of input data) in about 60 seconds. 

The Simulator will also be useful to do performance profiling of the whole OpenEMS Stack.

The Simulator App itself works, but not every planned function is fully implemented yet. Those ToDos are noted in the Pull-Request #1087 and the original branch is kept open for now.

Togehter with the Simulator App some more improvements and bug fixes were made in this Pull Request:
* Move global Cycle-Time to the "Core.Cycle" Component.
** Before there was an individual Cycle-Time per Scheduler, which proved to be not necessary and over-complicated things
** Global OpenEMS Cycle-Time can now adjusted in the "Core Cycle" Component in Apache Felix and OpenEMS UI
* Create and update Singleton Component Configurations via OpenEMS UI
** Improve detection + handling of Singletons in ComponentManager and EdgeConfig
* Fix Backend-Bug that would cause it to send all data all the time
* SymmetricEss: add ActivePower-Channel to Modbus Protocol
* UI: fix text-wrap in install component view