Fix behaviour for optional configuration parameters
An optional configuration parameter is a parameter that is defined as 'required = false' within the @AttributeDefinition in Config.java.
Before the behaviour for optional 'int' parameters used to be, that they would default to '0' if they were not provided. Now - maybe caused by an update to Apache Felix Configuration Admin - they default to an empty String which causes a NumberFormatException.