Fix controller behaviour in case of unavailable values.
This commit replaces `orElse()` method calls with `getOrError()`.
In case of unavailable (UNDEFINED) values, e.g. because an external meter is not available, this causes the Controller to throw an Exception instead of assuming the value is zero.
An exception in a Controller sets the `RunFailed` Fault-channel.
Specifically for control of energy storage systems: if no set-point is given, charge/dischare with zero Watt is set as default.

Also fix some Unit tests