Implement a Prediction architecture (#1327)
This new Predictor architecture uses modular `Predictors`, e.g. implementation of the `Predictor24Hours`. These Predictors can announce for which channels they allow a prediction and provide a `Prediction24Hours` result. 

The entire process of getting a prediction is handled by the `PredictorManager`. It knows which actual Predictors to ask for their prediction, post-processes these predictions and returns them either directly in code via the `Predictor::get24HoursPrediction()` method or via a `get24HoursPrediction` JSON-RPC request to the `_predictorManager` component.

The old prediction implementation is deprecated, but still kept in the code. It will be replaced by a follow-up pull-request.

Additionally:
- Add DummyTimedata implementation
- RRD4j: implement post-processing for required resolution
- SimulatorApp: fix Timedata service implementation