Implement MQTT (#1395)
This adds support for the MQTT protocol in OpenEMS Edge. The new "Controller.Api.MQTT" publishes data to an MQTT broker.

It uses the following MQTT topics:
- `edge/{OpenEMS-Edge-ID}/channel/{Component-ID}/{Channel-ID}`
  - Data of Channels is published on change and at least every 5 minutes to these topics.
  - Example: `edge/edge0/channel/_sum/GridActivePower`

- `edge/{OpenEMS-Edge-ID}/channel/lastUpdate`
  - Timestamp of the latest data update on any of the Channel topics above.
  - Example: `edge/edge0/channel/lastUpdate`

- `edge/{OpenEMS-Edge-ID}/edgeConfig`
  - Configuration of the OpenEMS Edge instance is published to this topic as a JsonObject.
  - Example: `edge/edge0/edgeConfig`

The implementation is built on version 5 of the MQTT protocol and was tested with Eclipse Mosquitto (https://mosquitto.org). 

Be aware that this implementation only marks the start of MQTT support in OpenEMS. Therefor implementation details might change with future versions.