Improve handling of OSGi References and target filters (#1603)
* Try to restart Components if they are unavailable

Due to a race condition on startup, sometimes Components are unavailable to the ComponentManager. The related error message is `[Core.ComponentManager(1)] : Could not get service from ref`. We are fixing this by restarting (i.e. 'reconfiguring') the Component in such a case.

* Increase ComponentManager 'ConfigNotActivated' level to 'FAULT'

* On UpdateComponent via ComponentManager: reset target properties

Before if a Component was referencing 'modbus0' and you would update it to 'modbus1', there were two possible behaviours. (1) if 'modbus0' was still existing, the update would work and the target filter would be set properly. (2) otherwise the target filter could not be set, because the Component would not reactivate. This would leave the Component in an error state. We are fixing this by always resetting the target properties.

* Set non-empty default target filter by default

If the target filter is made available via 'Config.java' and the configuration property is left as an empty string, the error `Invalid syntax in target property for dependency` arises on activate, which showed to also cause wrong behaviour later. We fix this by setting a default target filter of `(enabled=true)` - which is always true, but avoids the error

Thanks @sebastianasen for your valuable inputs.