Installation assistant: Reading serial numbers from devices is possible now + small adjustments (#95)
This pull request implements the possibility for serial numbers to be read in the corresponding serial number view. It is still possible to enter them manually if they can't be read.

Small changes in wording were also done in some places.

Last but not least an unused and unnecessary view was removed from the code.

Co-authored-by: sebastianasen <sebastian.asen@fenecon.de>
Co-authored-by: Stefan Feilmeier <stefan.feilmeier@fenecon.de>
Reviewed-on: https://git.intranet.fenecon.de/FENECON/fems/pulls/95
Reviewed-by: Stefan Feilmeier <stefan.feilmeier@fenecon.de>
Co-authored-by: Raphael Piller <raphael.piller@fenecon.de>
Co-committed-by: Raphael Piller <raphael.piller@fenecon.de>