Fix compatibility with Apache WebConsole 4.7.0 (#1618)
Latest version of Apache WebConsole does not explicitely store unchanged default configuration properties. This has certain implications on the way OpenEMS uses properties, e.g. it means that a Component that is created via Apache WebConsole with default 'id' does not have the 'id' stored as property.

See this issue for details:
https://issues.apache.org/jira/browse/FELIX-6436?focusedCommentId=17412472&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-17412472

* Read default property values from Factory
* Implement custom listConfigurations(componentId) method: this method also checks for the Component default value of the 'id' property
* Fix handling of static Component Property ID for singletons

Offtopic
* AbstractWebsocketServer: test for isShutdown