feature/GridOptimizedChargeRamp (#134)
Add a ramp with the ramp filter after the end time has been reached to gently remove the maximum charge restriction over a period of about five minutes.

Add a separate Channel for the original delay charge limit - this is used to calculate the real limit as an average of the last 100.

Update the tests

Co-authored-by: Maximilian Lang <maximilian.lang@fenecon.de>
Co-authored-by: Lukas Rieger <lukas.rieger@fenecon.de>
Co-authored-by: sebastianasen <sebastian.asen@fenecon.de>
Reviewed-on: https://git.intranet.fenecon.de/FENECON/fems/pulls/134
Reviewed-by: Stefan Feilmeier <stefan.feilmeier@fenecon.de>
Co-authored-by: sebastian asen <sebastian.asen@fenecon.de>
Co-committed-by: sebastian asen <sebastian.asen@fenecon.de>