Controller Dynamic Discharge (Time of use Pricing in Germany ) (#33)
Contains Dynamic discharge controller, which Schedules the discharging of an AC or DC energy storage system in the night based on Hourly-Prices, predicted production and consumption.

Contains the implementation according to the latest Predictor implementation.

Test cases are yet to be added followed by implementation of general API service for Time of Use pricing variants.

Co-authored-by: Sagar Venu <sagar.venu@fenecon.de>
Co-authored-by: Stefan Feilmeier <stefan.feilmeier@fenecon.de>
Reviewed-on: https://git.intranet.fenecon.de/FENECON/fems/pulls/33
Reviewed-by: Stefan Feilmeier <stefan.feilmeier@fenecon.de>
Co-authored-by: Sagar Venu <sagar.venu@fenecon.de>
Co-committed-by: Sagar Venu <sagar.venu@fenecon.de>