Hardy Barth Salia: implementation improvements (#142)
Small things like heartbeat were changed in the hardy barth software especially for the newer salia cpµ pro 2 (single charger).
Another problem was the external meter that is used internally in the hardy barth to provide the current charge values.
This meter also measures the standby consumption of the salia itself and is disabled by default - The function to enable it automatically is on hold for now, as we need a proper update process first.

Co-authored-by: sebastianasen <sebastian.asen@fenecon.de>
Reviewed-on: https://git.intranet.fenecon.de/FENECON/fems/pulls/142
Reviewed-by: Stefan Feilmeier <stefan.feilmeier@fenecon.de>
Co-authored-by: sebastian asen <sebastian.asen@fenecon.de>
Co-committed-by: sebastian asen <sebastian.asen@fenecon.de>