TimeOfUseTariff: UI Widgets + code fixes (#152)
Edge:
Added states functionality to the controller to display in UI.
Added modes for the controller with automatic and off functionality.
Added Active time counter to calculate the buying from grid state.

UI:
Added live widget for the controller to view the details in the online monitoring.
Added history widget for the controller to view the overall functionality of the controller and the prices and the soc curves.

Co-authored-by: venu-sagar <sagar.venu@fenecon.de>
Co-authored-by: Stefan Feilmeier <stefan.feilmeier@fenecon.de>
Reviewed-on: https://git.intranet.fenecon.de/FENECON/fems/pulls/152
Co-authored-by: Sagar Venu <sagar.venu@fenecon.de>
Co-committed-by: Sagar Venu <sagar.venu@fenecon.de>