Added classpath and project file to module, it complies and shows up in apache felix, but not in the eclipse workspace
This controller will oversee the Power of configured Evcs. Unlike the existing implementation, this will not need a battery.
It will limit the power based on a Power limitation and a limitation on the individual Phases. The limits can be configured static or based on connected power meters. That will also mean that it is possible to configure the Evcs Limiter to only use production energy from the system if the power limit is at 1W. It will calculate a „Resource Package“ based on the PowerLimit – Meter Value. This also means that the Evcs Limiter can be configured to take other devices on the power grid into account. So if a Power Limit of 30kW is configured and there are already other devices on the grid that consume 20kW of power, the Limiter will only allow the Evcs to consume 10kW.

In addition to that, the Limiter will also handle priority Evcs. Evcs that are configured to be priority, will not be reduced ore turned off unless there is no other possibility to stay under the limits (unbalanced loads are the only exception to this). The Evcs Limiter has a Priority Current configuration, that will determine how much current a Priority Evcs should have at a minimum, but if there is not enough power in the system to give that to all Priority Evcs, it will instead distribute the Priority current to all priority Evcs evenly.
As already stated, the EvcsLimiter can also be configured to check for an unbalanced load.



The EvscLimiter is designed to only need the ChargePower, PhaseCount and status of an evcs. Everything else will be handled by the Limiter, so the implementation of the Evcs can be as slim and „logic less“ as possible.
The Limiter has so far been successfully tested with: Schneider, Mennekes, Keba, Easee and Abl Evcs.