Changed name in project file
This implements the communication to the Easee Evcs over the proprietary Cloud api.
It uses a custom Bridge for this that communicates over http calls.
It will send the setPowerLimit channel content to the cloud, and retrieves the chargePower nad the Phases.
Since the Easee Cloud only allowes for 500 Calls every 5 minutes, it has a build in timer that will only allow the Evcs to update its values every 10 seconds.

The functionality of this component is not 100% guranteed, since a full test was not possible at the moment. Everything worked in an isolated environment with 3 Evcs, while only one of those where an Easee Evcs (the other where from Schneider and Mennekes).