Quickfix: remove translation from header (#390)
There is no quick solution for having working translation for "Live" and "History" tabs in header. Because of this, and because translations in all languages are currently very similar, texts are reverted to static for now.

When using `<ion-label translate>General.live</ion-label>` the translation is not applied and the text `GENERAL.LIVE` is shown in UI.

When using `<ion-label>{{ 'General.live' | translate }}</ion-label>` as per recommendation by Ionic docs, the ion-segment-buttons are not shown and clickable at all. In this case there is an exception in the logs: ERROR TypeError: i is undefined. I was not able to find the solution for now.

Reviewed-on: https://git.intranet.fenecon.de/FENECON/fems/pulls/390