Refactor Task Execution in Modbus Bridge
The idea is to try to execute all Write-Tasks as early as possible (directly after the EXECUTE_WRITE event) and all Read-Tasks as late as possible to have values available exactly when they are needed (i.e. at the BEFORE_PROCESS_IMAGE event).
A WaitHandler is introduced, that is responsible for calculating a 'wait' time before execution of Modbus ReadTasks.
It uses an internal queue of waiting times that would have been possible in previous Cycles. The minimum value of this queue is taken as next waiting time.

See discussion at OpenEMS Community: https://community.openems.io/t/zeitlicher-ablauf-der-modbus-tasks/734