UI: Switch back to static translate (instead of TranslateHttpLoader) (#455)
Sometimes TranslateHttpLoader does not provide translations early enough, so the translation key is shown instead of the actual translation. This bug is not reproducible, but happens often and is annoying.

This PR solves it by switching back to static loading of translations, while still keeping the i18n-ally compatible JSON format.

It also drops the App-Initializer because it is not anymore required (and did not solve the issue).

Reviewed-on: https://git.intranet.fenecon.de/FENECON/fems/pulls/455
Reviewed-by: Sagar Venu <sagar.venu@fenecon.de>