AppManager: fix removing ips when deinstalling app (#457)
Before if two apps needed the same ip in the network configuration and one got deinstalled the ip got removed as well. Now the ip only gets removed if both apps get deinstalled.

Also added possability to specify on which interface the ip should get added and possability to set label for ip e. g. 'Evcs' for Evcs Apps to know where the ip belongs to.

Co-authored-by: Stefan Feilmeier <stefan.feilmeier@fenecon.de>
Reviewed-on: https://git.intranet.fenecon.de/FENECON/fems/pulls/457
Reviewed-by: Hueseyin Sahutoglu <hueseyin.sahutoglu@fenecon.de>
Reviewed-by: Stefan Feilmeier <stefan.feilmeier@fenecon.de>
Co-authored-by: Michael Grill <michael.grill@fenecon.de>
Co-committed-by: Michael Grill <michael.grill@fenecon.de>