Add OpenEMS Contribution Guidelines (#2042)
Put together the contribution guidelines from @fabianfnc and sfeilmeier and added thoughts from consolinno and opernikus developers to it.

See discussion at https://community.openems.io/t/openems-code-contribution-guidelines/1182/3

I am merging this as a first version, which we can continuously build on. Specifically all the bullet points should be explained with examples. But it's better to have something than nothing for now.

Co-authored-by: Christian Lehne <c.lehne@opernikus.de>
Co-authored-by: Stefan Feilmeier <stefan.feilmeier@fenecon.de>