Implement the Alpitronic Hypercharger DC charging station (#2071)
Implementation of the Alpitronic Hypercharger V1-X DC electric vehicle charging station.

The charging station comes with up to four physical connectors. This implementation is designed in a way, that each connector is represented by an individual OpenEMS Component. So if you have a Alpitronic with two connectors, please configure two "EVCS Alpitronic Hypercharger" with the connector configuration property set to SLOT_0 and SLOT_1 respectively.

Supported Firmware versions: 1.7.2 and higher.

Co-authored-by: Lorant Meszlenyi <16509320+hydroid7@users.noreply.github.com>
Co-authored-by: Sebastian Asen <47855186+sebastianasen@users.noreply.github.com>