Channel: never reset cumulated Channels to null
Per definition, the value of a cumulated Channel must be steadily increasing. This PR makes sure, that the value never gets reset to 'null'.

This avoids the issue, that when communication to a device (e.g. a PV-Inverter) gets lost during the night, the energy values get reset to 'null' and as a result the calculation of energy in the Sum component is broken.