Update GenericHttpWorker.java
With this version a total of 3 bugs have been fixed.

Bug 1:
delete_task did not delete URL's because the check if the URL is present in the URL array was incorrect.

Bug 2:
push_task always added every URL even if the URL was already in the URL array. This was due to an incorrect check if the URL already exists in the array.

Bug 3:
In the forever() function in the part where the external URL's are processed (via push_task) an ArrayIndexOutOfBoundsError could occur. The bug was discovered after bug 1 was fixed. This bug is due to the instant deletion of an element from the URL array after it has been processed.
E.g. the String[] holds two elements:
After one run the this.delete_task() function is called. This function modifies the this.add_external_task to reduce the number of elements to one. In this case before the this.delete_task() function is called there is an Index 1 after this only Index 0 exists. However, if j is used as a variable for the index , j is increased to 1 after the first run of the for-Loop.