GoodWe.BatteryInverter: use data from battery directly (#689)
In fems888 I experienced, that the `batteryInverter0/BmsChargeMaxCurrent` was `25 A` instead of `0 A` for a full battery. This caused the calculation of surplus power to be wrong and the inverter did not switch to AUTO mode.

Effectively most of the times, the battery inverter did still switch to AUTO mode in the end and was behaving correctly, but
1. there were seldom cases, where because of this the battery did not discharge
2. in Grafana the "SetActivePower" was very different to the actual "ActivePower"

This PR uses the battery data directly and calculates the surplus power correctly - fixing both issues above.

Reviewed-on: https://git.intranet.fenecon.de/FENECON/fems/pulls/689
Reviewed-by: Sebastian Asen <sebastian.asen@fenecon.de>