Modbus Bridge: refactor & cleanup Elements (#2273)
This is a follow-up PR to #1976. 

The initial intention for this PR was a bug that existed previously: When the Modbus Bridge detects a 'defective device', i.e. it is not possible to read from an external device connected via modbus, e.g. because it is turned off, somebody unplugged the cable, etc., all Channels are 'invalidated'. This invalidation did not work properly for individual Bits in `BitsWordElements`.

When looking at the previous code it becomes obvious how such a bug could happen. The structure and inheritance was complicated and @Overrides were often unclear. Also there were hardly any JUnit tests to validate the function.

Because of this I had added JUnit tests in #2283, then refactored all Modbus Element classes and revalidated that they behave exactly the same way as before with the JUnit tests.

This implementation is tested on a couple of production systems, including my own private system.