FENECON Home: fix switching Battery Start Up Relay ON
This fixes a bug, that keeps a Home Battery from getting started via the Battery Start Up Relay. The `ModbusCommunicationFailed` State has level FAULT. Because of this `battery.hasFaults()` is never true when the battery is turned off, so the State-Machine immediately goes to ERROR state and never enters GO_RUNNING when the `ModbusCommunicationFailed` State is set. It still works in many cases after a full cold-start (e.g. by restarting OpenEMS service), though, because then `ModbusCommunicationFailed` State is not set _yet_.

This fix handles `ModbusCommunicationFailed` State separately in UNDEFINED state and waits for modbus communication in GO_RUNNING. In case of an overall timeout - i.e. the battery cannot be started - `MaxStartAttemptsFailed` is still triggered.

In the process of refactoring the GO_RUNNING handler, I implemented a full Sub-State-Machine including nice debug log messages and a Mermaid chart.

I also touched the OpenEMS testing framework to force set a Channel (to cirumvent channel debounce settings)

Co-authored-by: Sebastian Asen <47855186+sebastianasen@users.noreply.github.com>
Reviewed-by: Hueseyin Sahutoglu <34771592+huseyinsaht@users.noreply.github.com>