FEMS Backport 2023-09-25 (#2368)
* SolarLog: fix setting/calculation of active production energy

Added missing event subscription for "TOPIC_CYCLE_AFTER_PROCESS_IMAGE" to calculate ActiveProductionEnergy & removed reading this value via modbus

* App Center: Refactor App.Api.ModbusTcp

* ElementToChannelConverter:  Add SET_NULL_FOR_DEFAULT

* GoodWe BatteryInverter: validate compatible FENECON Home Battery

- Check relation between goodwe inverter and fenecon home battery
- Calculate Number of Modules for GoodWe depending on the battery values

* AppCenter: Home 20 & 30

Implementation of Apps for Home 20 & 30

* GoodWe BatteryInverter: improve Warning/Error State descriptions

- Fixes mapping of Registers and bits.
- New format of the description:
[Fault English | Fault German | Suggested solution German]

* KACO: fit ActivePower set point percentage -100 to 100

The filter is from -100 to +100 factor.

* FENECON Home: fix switching Battery Start Up Relay ON

This fixes a bug, that keeps a Home Battery from getting started via the Battery Start Up Relay. The `ModbusCommunicationFailed` State has level FAULT. Because of this `battery.hasFaults()` is never true when the battery is turned off, so the State-Machine immediately goes to ERROR state and never enters GO_RUNNING when the `ModbusCommunicationFailed` State is set. It still works in many cases after a full cold-start (e.g. by restarting OpenEMS service), though, because then `ModbusCommunicationFailed` State is not set _yet_.

This fix handles `ModbusCommunicationFailed` State separately in UNDEFINED state and waits for modbus communication in GO_RUNNING. In case of an overall timeout - i.e. the battery cannot be started - `MaxStartAttemptsFailed` is still triggered.

In the process of refactoring the GO_RUNNING handler, I implemented a full Sub-State-Machine including nice debug log messages and a Mermaid chart.

I also touched the OpenEMS testing framework to force set a Channel (to cirumvent channel debounce settings)

* Modbus Bridge: hide error logs on first try (on configured low LogVerbosity)

This avoids unnecessary logs and resemble the logging behavour of the previous Modbus bridge

* UI: Fix notch color for all OEM's

Replace notch color with OEM's background-color

* App Center: return created dependencies

Necessarry for selecting AC-Meter in Home IBN and update it.

* Meter SDM630: calculate active production/consumption energy manually

* Battery: standardize debugLog

* UI: Splitting index into login and overview

* Solves Community Issue [OpenEMS](https://community.openems.io/t/setup-ide-for-openems-ui-not-working/1923/5?u=stefan.feilmeier)
* Due to lifecycle issues and routing issues after login, the `login` and the `edges-overview` get split into two components.
  * If only one `edge` assigned to the user -> the user logs in, the overview wont be shown, but because this is happening in the same component, routing forward creates a problem in `live`. The currentData is subscribed but not shown. Only on reload it worked again.
* This also fits the purpose of reducing unnecessary requests.

---------

Co-authored-by: Michael Grill <59126309+michaelgrill@users.noreply.github.com>
Co-authored-by: Hueseyin Sahutoglu <34771592+huseyinsaht@users.noreply.github.com>
Co-authored-by: Sebastian Asen <47855186+sebastianasen@users.noreply.github.com>
Co-authored-by: Pooran Chandrashekaraiah <46567310+pooran-c@users.noreply.github.com>
Co-authored-by: Anas Shetla <141644226+AnasShetla@users.noreply.github.com>
Co-authored-by: Mohammadmahdi Ataei <54065538+mahdiataie@users.noreply.github.com>