Bugfix UIWebsocket does not open WebsocketServer on missed  Event
The UI Websocketserver will not call "startServer" on a missed AFTER_IS_INITIALIZED Event.

This happens, when Metadata is initialized, before the WebsocketServer gets configured. (This can happen on restart of the backend too)
Therefore the activate Method in UI Websocket has to check the metadata if it is initialized and start the Server.