added AlertingSetting Entry to Field
This doesn't do anything yet. But will in the future, when AlertingSettings get refactored to UserAlertingSettings and OdooHandler calls different db table.
This commit is one of many prepare PRs / commits for a bigger PR:
SumStateAlert #2260
By Seperating the big #2260 into smaller commits we would like to make the review easier.

Co-authored-by:  Kai Jeschek <99220919+da-Kai@users.noreply.github.com>