Refactored Name of AlertingSetting to UserAlertingSettings
This is just a refactor of the Name. This does not provide any functionality, but will improve the readability of the SumStateAlert PR in the future.
This commit is one of many prepare PRs / commits for a bigger PR:
SumStateAlert #2260
By Seperating the big #2260 into smaller commits we would like to make the review easier.

Co-authored-by:  Kai Jeschek <99220919+da-Kai@users.noreply.github.com>