Zone selection config for Time-of-Use-Tariff Awattar (#1009)
Time-of-use-Tariff Awattar currently only offers price information for Germany. However, for users in Austria, a separate URL needs to be incorporated.
In this implementation the system is updated by introducing a country selection option in the configuration. Users can now choose between Germany and Austria, and the system will provide country-specific prices accordingly.

Co-authored-by: Stefan Feilmeier <stefan.feilmeier@fenecon.de>
Reviewed-on: https://git.intranet.fenecon.de/FENECON/fems/pulls/1009
Reviewed-by: Hueseyin Sahutoglu <hueseyin.sahutoglu@fenecon.de>
Co-authored-by: Sagar Bandi Venu <sagar.venu@fenecon.de>
Co-committed-by: Sagar Bandi Venu <sagar.venu@fenecon.de>