ToU: calculate essMaxChargePower from usable Capacity (#1010)
This improves the charging behaviour with small capacity batteries. Before the charge power was configurable (but in fact hard-coded to 5 kW) and mainly tested on a 22 kWh system (fems888). With this PR the power in CHARGE mode is calculated with a C-Rate of 0.25 i.e. power is 1/4 of the (usable) capacity. For a 22 kWh system it would still be 5.5 kW, but for a 8.8 kWh battery its 2.2 kW then. It should also scale better with Home 20 & 30, i.e. larger batteries.

Reviewed-on: https://git.intranet.fenecon.de/FENECON/fems/pulls/1010