KACO Battery Inverter: quicker control behaviour (#1032)
I added the setting of the "WparamRmpTms" param to 100 ms.
The WparamRmpTms parameter constrains performance changes using a PT1 behavior.
By default, a 1-second (1000 ms) duration is stored here. This duration can be reduced to 0.1 seconds (100 ms) for quicker control behavior.
While a complete reduction to 0 is technically possible, it may result in overcurrent or overvoltage events, especially in high-power changes and multiple devices.
This feature benefits FFR use cases and aids in preventing battery derating.

Reviewed-on: https://git.intranet.fenecon.de/FENECON/fems/pulls/1032
Reviewed-by: Hueseyin Sahutoglu <hueseyin.sahutoglu@fenecon.de>
Reviewed-by: Stefan Feilmeier <stefan.feilmeier@fenecon.de>
Co-authored-by: Pooran Chandrashekaraiah <pooran.c@fenecon.de>
Co-committed-by: Pooran Chandrashekaraiah <pooran.c@fenecon.de>