Zone selection config for Time-of-Use-Tariff Awattar
Time-of-use-Tariff Awattar currently only offers price information for Germany. However, for users in Austria, a separate URL needs to be incorporated.
In this implementation the system is updated by introducing a country selection option in the configuration. Users can now choose between Germany and Austria, and the system will provide country-specific prices accordingly.

Co-authored-by: Sagar Venu <32655208+venu-sagar@users.noreply.github.com>
Co-authored-by: Hueseyin Sahutoglu <34771592+huseyinsaht@users.noreply.github.com>